refactor: 서버 매핑 모델의 불필요한 프로퍼티 제거
